users/serializers.py
from rest_framework import serializers
from .models import User
from dj_rest_auth.registration.serializers import RegisterSerializer

class SignupSerializer(RegisterSerializer):
    nickname = serializers.CharField(max_length=100)
    _has_phone_field = False
    
    def get_cleaned_data(self):
        data = super().get_cleaned_data()
        data['nickname'] = self.validated_data.get('nickname', '')
        return data
    
    def save(self, request):
        user = super().save(request)
        user.nickname = self.validated_data.get('nickname', '')
        user.save()
        return user

class ProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['username', 'nickname', 'email']
        read_only_fields = ['email']
