requirements.txt
# Django 기본
Django==5.2.4

# Django REST Framework
djangorestframework==3.15.2

# JWT 인증
djangorestframework-simplejwt==5.3.0
dj-rest-auth[with_social]==6.0.0

# Django Allauth (소셜 로그인)
django-allauth==0.63.3

# CORS 처리
django-cors-headers==4.3.1

# HTTP 요청 (seed_movies.py에서 사용)
requests==2.31.0

# 환경변수 관리 (권장)
python-decouple==3.8

# 데이터베이스 (SQLite는 Python 내장이지만 PostgreSQL 사용시)
psycopg2-binary==2.9.9

# 보안 강화 패키지들 (선택사항)
django-ratelimit==4.1.0
django-defender==0.9.7

# 개발환경 도구들 (선택사항)
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# 테스트 도구 (선택사항)
pytest-django==4.8.0
factory-boy==3.3.0

# 배포 관련 (선택사항)
gunicorn==21.2.0
whitenoise==6.6.0